requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
tabula-py>=2.9.0
jpype1>=1.4.0
psycopg2-binary>=2.9.0
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.1